Split server in different files
